move appsec channels file
