+++ packages/dd-trace/src/appsec/channels.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const dc = require('../../../../diagnostics_channel')
-// TODO move this file to another path, or delete and use channel directly
+
 // TODO: use TBD naming convention
 //       or directly use http plugin's channels
 //       when it gets converted to new plugin system
