+++ packages/dd-trace/src/appsec/index.js
@@ -4,7 +4,7 @@ const log = require('../log')
 const RuleManager = require('./rule_manager')
 const waf = require('./waf')
 const remoteConfig = require('./remote_config')
-const { incomingHttpRequestStart, incomingHttpRequestEnd } = require('./gateway/channels')
+const { incomingHttpRequestStart, incomingHttpRequestEnd } = require('./channels')
 const addresses = require('./addresses')
 const Reporter = require('./reporter')
 const web = require('../plugins/util/web')
