+++ packages/dd-trace/test/appsec/index.spec.js
@@ -8,7 +8,7 @@ const WAFContextWrapper = require('../../src/appsec/waf/waf_context_wrapper')
 const RuleManager = require('../../src/appsec/rule_manager')
 const remoteConfig = require('../../src/appsec/remote_config')
 const appsec = require('../../src/appsec')
-const { incomingHttpRequestStart, incomingHttpRequestEnd } = require('../../src/appsec/gateway/channels')
+const { incomingHttpRequestStart, incomingHttpRequestEnd } = require('../../src/appsec/channels')
 const Reporter = require('../../src/appsec/reporter')
 const agent = require('../plugins/agent')
 const Config = require('../../src/config')
